Tidied up allEvents file. Sorting through the imported calendar data is still needed.
